# PERF: Fast path for dtype lookup of float and long