--- a/numpy/core/src/multiarray/array_coercion.c
+++ b/numpy/core/src/multiarray/array_coercion.c
@@ -216,9 +216,9 @@ npy_discover_dtype_from_pytype(PyTypeObject *pytype)
         DType = Py_None;
     }
     else if (pytype == &PyFloat_Type) {
-        DType = &PyArray_PyFloatAbstractDType;
+        DType = (PyObject *)&PyArray_PyFloatAbstractDType;
     } else if (pytype == &PyLong_Type) {
-        DType = &PyArray_PyIntAbstractDType;
+        DType = (PyObject *)&PyArray_PyIntAbstractDType;
     }
     else {
         DType = PyDict_GetItem(_global_pytype_to_type_dict,