@seberg 2022-05-18
At some point I was wondering if tiny LRU of recently seen types can work (even within the same call for recursive cases), but with the plausibility of subinterpreters, I am not sure it is fun to spend much time on figuring it out.

Anyway, I am happy with adding a fast path for ints and floats here (still some code nitpicks).  I am curious how much it makes the benchmarks faster (I think thre should be one be some benchmarks hitting this for `bench_array_coercion`).