@eendebakpt 2022-05-18
> At some point I was wondering if tiny LRU of recently seen types can work (even within the same call for recursive cases), but with the plausibility of subinterpreters, I am not sure it is fun to spend much time on figuring it out.
> 
> Anyway, I am happy with adding a fast path for ints and floats here (still some code nitpicks). I am curious how much it makes the benchmarks faster (I think thre should be one be some benchmarks hitting this for `bench_array_coercion`).

Improvement on the benchmarks is not a lot. From profiling results, the improvement on `np.sqrt(1.1)` is at best 1% to 2% (and similar for most ufuncs on a float). The expensive part is the call to `PyDict_GetItem` on `_global_pytype_to_type_dict`. Instead of a fast path we could use a structure that that is more efficient than the python dictionary (that one has to calculate hashes for the objects), something like a `std::map<PyObject *, PyObject *>`. We then skip any hash calculations, and if works for all types. (I am thinking of `np.float64`). The new structure would either have to replace the `_global_pytype_to_type_dict`, or act as a fast path with the `_global_pytype_to_type_dict` as a fallback.

Even though it is not a lot of work, 1% gain is not a lot so I am considering closing this PR.