@seberg 2022-05-18
Ah OK, maybe that is why I didn't add it.  The case that I thought might gain quite a bit would be:
```
arr = np.array([1000.] * 10000)
```
but, I guess they should also gain quite well to be worthwhile.