@eendebakpt 2022-05-19
> Ah OK, maybe that is why I didn't add it. The case that I thought might gain quite a bit would be:
> 
> ```
> arr = np.array([1000.] * 10000)
> ```
> 
> but, I guess they should also gain quite well to be worthwhile.

For the example you gave performance improvement is about 20-25%. But for a smaller list the gain will be less. 
For this particular example a LRU with just 1 element would probably work, or the current fast path for `long` and `float`. 
If a more general approach covering all types is just as fast I would still prefer that.