@seberg 2022-05-19
The problem with an LRU is that I am not sure it is wise to make it persistent over multiple calls,  `type` objects are not necessarily immortal (although the ones that have DTypes effectively are right now).
OTOH, it could still be very effective if it is persistent only in a single (recrusive) call for those large lists, that is.

So, I am happy for you to look into it if it interests you, but I am not sure it is particularly high priority. (priority is very subjective though :))