@eendebakpt 2022-05-19
> The problem with an LRU is that I am not sure it is wise to make it persistent over multiple calls, `type` objects are not necessarily immortal (although the ones that have DTypes effectively are right now). OTOH, it could still be very effective if it is persistent only in a single (recrusive) call for those large lists, that is.
> 
> So, I am happy for you to look into it if it interests you, but I am not sure it is particularly high priority. (priority is very subjective though :))

I agree priority is not so high, but since I am working on it I might complete it anyway.

Besides this PR, I tested two more variations:

1) This https://github.com/numpy/numpy/compare/main...eendebakpt:dtype_lookup_v2a is an implementation of the lookup table approach I had in mind. The advantage is that it works for all types, not just a few selected ones.
2) https://github.com/numpy/numpy/compare/main...eendebakpt:dtype_lookup_v3 is a single-element LRU. Advantage it is adaptive (so large lists are covered, whether they are `int` or `float` or something else) a small addition. Downside is that mixed cases are not handled as good as by 1)

Performance numbers on `np.array([0] * 100)`:
```
main: 1.21 seconds
this PR: 1.0 seconds
LUT approach: between 1.08 and 1.12 seconds
LRU: 0.97 seconds
```
The results of the LUT approach depends a bit on what is in the lookup table. In my tests adding more types (e.g. `np.int16`) could improve performance when the type searched for (e.g. `int`) is easier to find with the binary search.

@seberg Based on the above I have a small preference for either the LRU approach or the approach in this PR. 

<details>
<summary>dtype_lookup_v3: python runtests.py --bench-compare main bench_array_coercion</summary>

```
       before           after         ratio
     [57961d23]       [6be51bac]
     <main>           <dtype_lookup_v3>
-        546±20ns        519±0.5ns     0.95  bench_array_coercion.ArrayCoercionSmall.time_array_dtype_not_kwargs(array([5]))
-        470±10ns          447±1ns     0.95  bench_array_coercion.ArrayCoercionSmall.time_ascontiguousarray(1)
-         460±6ns          437±4ns     0.95  bench_array_coercion.ArrayCoercionSmall.time_array_no_copy(1)
-         387±8ns          368±5ns     0.95  bench_array_coercion.ArrayCoercionSmall.time_array(5)
-       399±0.9ns          379±2ns     0.95  bench_array_coercion.ArrayCoercionSmall.time_asanyarray(1)
-         454±5ns          429±1ns     0.94  bench_array_coercion.ArrayCoercionSmall.time_ascontiguousarray(5)
-         390±6ns          366±2ns     0.94  bench_array_coercion.ArrayCoercionSmall.time_asanyarray(5)
-        420±10ns          382±2ns     0.91  bench_array_coercion.ArrayCoercionSmall.time_array(1)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```
</details>

<details>
<summary>dtype_lookup_v3: python runtests.py --bench-compare main bench_core</summary>

```
      before           after         ratio
     [57961d23]       [6be51bac]
     <main>           <dtype_lookup_v3>
+      67.1±0.3μs       82.8±0.3μs     1.23  bench_core.PackBits.time_packbits_little(<class 'numpy.uint64'>)
+        995±10μs      1.09±0.01ms     1.09  bench_core.PackBits.time_packbits_axis1(<class 'numpy.uint64'>)
+         169±4μs          182±7μs     1.08  bench_core.CountNonzero.time_count_nonzero(1, 1000000, <class 'numpy.int32'>)
+      51.0±0.5μs       54.9±0.3μs     1.08  bench_core.PackBits.time_packbits(<class 'numpy.uint64'>)
+     1.07±0.01ms      1.15±0.03ms     1.07  bench_core.PackBits.time_packbits_axis0(<class 'numpy.uint64'>)
+         147±1μs          157±1μs     1.07  bench_core.CountNonzero.time_count_nonzero(3, 1000000, <class 'numpy.int8'>)
-         345±4ns          327±2ns     0.95  bench_core.Core.time_array_empty
-         100±1μs       95.1±0.7μs     0.95  bench_core.UnpackBits.time_unpackbits_axis1
-      39.3±0.6μs       36.9±0.3μs     0.94  bench_core.CorrConv.time_convolve(50, 1000, 'valid')
-      10.3±0.3μs       9.51±0.2μs     0.93  bench_core.CountNonzero.time_count_nonzero_axis(3, 100, <class 'str'>)
-     1.99±0.06μs      1.84±0.01μs     0.93  bench_core.Core.time_array_l_view
-         455±2μs          417±2μs     0.92  bench_core.CountNonzero.time_count_nonzero_multi_axis(2, 10000, <class 'object'>)
-     6.14±0.04μs      5.62±0.05μs     0.92  bench_core.UnpackBits.time_unpackbits
-     44.6±0.06ms       40.8±0.1ms     0.91  bench_core.CountNonzero.time_count_nonzero_multi_axis(2, 1000000, <class 'object'>)
-         291±4μs          265±2μs     0.91  bench_core.PackBits.time_packbits_axis0(<class 'bool'>)
-         684±4μs          622±2μs     0.91  bench_core.CountNonzero.time_count_nonzero_multi_axis(3, 10000, <class 'object'>)
-      67.3±0.4ms       61.1±0.1ms     0.91  bench_core.CountNonzero.time_count_nonzero_axis(3, 1000000, <class 'object'>)
-         682±3μs          618±2μs     0.91  bench_core.CountNonzero.time_count_nonzero_axis(3, 10000, <class 'object'>)
-      17.8±0.8μs      16.1±0.09μs     0.91  bench_core.Core.time_diagflat_l50_l50
-      44.9±0.4ms       40.7±0.1ms     0.91  bench_core.CountNonzero.time_count_nonzero_axis(2, 1000000, <class 'object'>)
-         234±3μs        211±0.6μs     0.90  bench_core.CountNonzero.time_count_nonzero_axis(1, 10000, <class 'object'>)
-     22.7±0.05ms      20.5±0.05ms     0.90  bench_core.CountNonzero.time_count_nonzero_axis(1, 1000000, <class 'object'>)
-         464±3μs          418±2μs     0.90  bench_core.CountNonzero.time_count_nonzero_axis(2, 10000, <class 'object'>)
-      67.9±0.6ms      61.0±0.06ms     0.90  bench_core.CountNonzero.time_count_nonzero_multi_axis(3, 1000000, <class 'object'>)
-        16.7±1μs       14.9±0.1μs     0.89  bench_core.Core.time_diagflat_l100
-        327±20ns          290±5ns     0.88  bench_core.Core.time_array_1
-      6.99±0.2μs      5.76±0.02μs     0.82  bench_core.Core.time_array_l100
-      13.4±0.1μs      10.9±0.05μs     0.82  bench_core.Core.time_diag_l100
-      49.9±0.4μs       40.5±0.2μs     0.81  bench_core.Core.time_array_float_l1000
-      51.8±0.1μs       41.7±0.2μs     0.81  bench_core.Core.time_array_int_l1000
-        39.5±2μs       28.4±0.2μs     0.72  bench_core.Core.time_array_float64_l1000

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```

(benchmark claims performance decreased, but I think this is just fluctuations on my system. overall there seems to be an increase)
</details>